requirements/common.txt
Django==2.2.5
django-cors-headers==3.0.2
django-extensions==2.2.1
djangorestframework==3.10.3
python-decouple==3.1
boto3==1.9.223
botocore==1.12.253
docutils==0.15.2
jmespath==0.9.4
pytz==2019.3
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
python-dateutil==2.8.0
psycopg2-binary-2.8.4
validate-email-1.3
django-multiselectfield-0.1.10